+++ yesod-form/ChangeLog.md
@@ -1,3 +1,8 @@
+## 1.6.2
+
+* Move `addClass` from private/undocumented in `Yesod.Form.Bootstrap3` to `Yesod.Form.Functions` [#1510](https://github.com/yesodweb/yesod/pull/1510)
+* Add `Yesod.Form.Functions.removeClass` [#1510](https://github.com/yesodweb/yesod/pull/1510)
+
 ## 1.6.1
 
 * Explicitly define `(<>)` in the `Semigroup` instance for `Enctype`